tester: add double HTTP endpoint PBT (clj) tester.
This loads the actions.ebnf parser grammar and uses it as the generator
specification with instacheck. The requests are encoded as JSON
containing method, path, payload). These requests are sent to
both endpoints and compared (essentially using one as an test oracle for
the other).

Run like this:
    clj -M -m tester.core ./actions.ebnf http://localhost:8000 http://localhost:8001

Or wrap it in a container (this should work on macos/colima too):
    docker run -it --network host -v $(pwd):$(pwd) -w $(pwd) -u $(id -u) -v $HOME:$HOME -e HOME=$HOME clojure:temurin-23-tools-deps-noble \
        -M -m tester.core ./actions.ebnf http://localhost:8000 http://localhost:8001