Buggy PUT triggered by mode and api IP/index.
Add a "buggy" module/mode. All the module does is set the BUGGY
environment variable.

In the app api server, wait for eth0 to be present with an IP. Then read
the IP and use the last octect as the server ID. If the server is the
second one (either "2" for static or "101" for DHCP), then trigger the
buggy code in the PUT endpoint that swpas the name and email.