tester: move reset to tester and add final query
Instead of generating the reset as part of the generated test definition
in the `actions.ebnf` file, add the reset action in the actual
run-actions function. This reduces noise when reporting the test case to
the user (the reset is always there so it's not interesting).

Also, add a query of all users (in the tester again) at the end of the
actions list to do the final comparison of the user state.