Return general result from POST and PUT.
Instead of returning the new or updated full user properties in the
response, just return a generic message. This allows bugs/errors to
linger until the final query which will cause the shrinking process to
to do more work/requests in order to get a new result (which might
actually be bigger due to the non-deterministic failure from the point
of the tester).