Add monitor mode: prometheus, grafana, blackbox-exporter.
Add prometheus configuration to use blackbox-exporter to ping the direct
and balanced,dhcp endpoints and visualize it from grafana.

This monitor mode is a separate mode that runs the containers in the
host network environment so that it can reach the published ports for
the direct and balanced,dhcp modes.

The prometheus probes and default grafana dashboard are configured to
query and show the data every second.