Add waits for message-bus and stream. Fix perms.
Add wait.sh on interfaces in balancer and message-bus. Also, in the
balancer, wait for the stream to actually appear. On faster system,
things happen faster and in different order so we need to actually wait.

Also, add newer capabilities needed for the balancer and api services to
be able to run.