tester: Use instacheck.cli 0.9.3. Add parse cmd.
Major refactoring to use instacheck.cli from instacheck 0.9.3 which
allows use of the cli routines externally in a more friendly way (no
usage, ability to pass a check-fn to do-check, etc).

Use the instacheck.cli "library" for the check command and the parse
command. Support a bunch of standard command line options (see the usage
docstring).

Add a `bin/tester` wrapper that invokes the clojure standard image to
run the deps.edn based tester command. Has the deps config setup so that
it runs from the top instead of inside the tester directory. Results
from checks will be in the `samples/` directory by default.

Improve URL handling so that a trailing slash in the destinations won't
break things.