presentation: compose/conlink/mdc content/diagrams.
Add intro, personal intro, compose, conlink, and mdc content start.

Add a simple-compose.yaml and use it for the compose section of the
presentation.

Switch from showing the drawio diagram directly in the presentation to
using exported pngs (the drawio viewer does not size the diagram well).
Move the main drawio diagram to top docs level.